Move find_candidates_in_parent_dirs() into a config module so that config doesn't import from cli.
